# doc/swagger.yaml
swagger: '2.0'
info:
  version: '1.0-SNAPSHOT'
  title: TC Service Identity
  license:
    name: MIT
    url: http://github.com/gruntjs/grunt/blob/master/LICENSE-MIT
host: api.topcoder.com
basePath: /v3
schemes:
- http
- https
securityDefinitions:
  bearer:
    type: apiKey
    name: Authorization
    in: header
paths:
...
  /roles/{roleId}:
    get:
      description: Fetch a particular role information
      tags:
      - roles
      operationId: RoleResource.getObject
      security:
        - bearer: []
      produces:
      - application/json
      parameters:
      - name: roleId
        in: path
        required: true
        type: string
        description: role id
      - name: fields
        in: query
        required: false
        type: string
        description: |
          fields=fieldName1,fieldName2,...,fieldN  - parameter for
          choosing which fields of role that will be included in response.

          + id
          + createdAt
          + createdBy
          + updatedAt
          + updatedBy
          + roleName
          + subjects
      responses:
        200:
          description: OK
          schema:
            type: object
            properties:
              id:
                type: string
              version:
                type: string
              result:
                type: object
                properties:
                  success:
                    type: boolean
                  status:
                    type: integer
                  metadata:
                    type: object
                  content:
                    $ref: "#/definitions/RoleResponse"
        400:
          description: Invalid input.
          schema:
            $ref: "#/definitions/ErrorModel"
        401:
          description: If user not authenticated.
          schema:
            $ref: "#/definitions/ErrorModel"
        404:
          description: If role does not exist.
          schema:
            $ref: "#/definitions/ErrorModel"
        500:
          description: Internal server error.
          schema:
            $ref: "#/definitions/ErrorModel"
    put:
      description: Update a particular role
      tags:
      - roles
      operationId: RoleResource.updateObject
      security:
        - bearer: []
      produces:
      - application/json
      parameters:
      - name: roleId
        in: path
        required: true
        type: string
        description: role id
      - in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/NewRoleBodyParam'
      responses:
        200:
          description: OK
          schema:
            type: object
            properties:
              id:
                type: string
              version:
                type: string
              result:
                type: object
                properties:
                  success:
                    type: boolean
                  status:
                    type: integer
                  metadata:
                    type: object
                  content:
                    $ref: "#/definitions/RoleResponse"
        400:
          description: Invalid input.
          schema:
            $ref: "#/definitions/ErrorModel"
        401:
          description: If user not authenticated.
          schema:
            $ref: "#/definitions/ErrorModel"
        403:
          description: If user does not have admin role.
          schema:
            $ref: "#/definitions/ErrorModel"
        404:
          description: If role does not exist.
          schema:
            $ref: "#/definitions/ErrorModel"
        500:
          description: Internal server error.
          schema:
            $ref: "#/definitions/ErrorModel"
    delete:
      description: Delete a role
      tags:
      - roles
      operationId: RoleResource.deleteObject
      security:
        - bearer: []
      produces:
      - application/json
      parameters:
      - name: roleId
        in: path
        required: true
        type: string
        description: role id
      responses:
        200:
          description: OK. Response will contain id of deleted role.
          schema:
            type: object
            properties:
              id:
                type: string
              version:
                type: string
              result:
                type: object
                properties:
                  success:
                    type: boolean
                  status:
                    type: integer
                  metadata:
                    type: object
                  content:
                    type: object
                    properties:
                      id:
                        type: string
        400:
          description: Invalid input.
          schema:
            $ref: "#/definitions/ErrorModel"
        401:
          description: If user not authenticated or does not have admin role.
          schema:
            $ref: "#/definitions/ErrorModel"
        403:
          description: If user does not have admin role.
          schema:
            $ref: "#/definitions/ErrorModel"
        404:
          description: If role does not exist.
          schema:
            $ref: "#/definitions/ErrorModel"
        500:
          description: Internal server error.
          schema:
            $ref: "#/definitions/ErrorModel"
  /roles/{roleId}/assign:
    post:
      description: Assign a role to a user
      tags:
      - roles
      operationId: RoleResource.assignRole
      security:
        - bearer: []
      produces:
      - application/json
      parameters:
      - name: roleId
        in: path
        required: true
        type: string
        description: role id
      - name: filter
        in: query
        required: false
        type: string
        description: |
          fieldName=fieldValue (enum[string],optional)  - parameter for
          filtering the fields according to v3 spec

          + id
          + createdAt
          + createdBy
          + updatedAt
          + updatedBy
          + roleName
          + subjectID
      - name: fields
        in: query
        required: false
        type: string
        description: |
          fields=fieldName1,fieldName2,...,fieldN  - parameter for
          choosing which fields of role that will be included in response.

          + id
          + createdAt
          + createdBy
          + updatedAt
          + updatedBy
          + roleName
          + subjects
      - $ref: "#/parameters/offsetParam"
      - $ref: "#/parameters/limitParam"
      - $ref: "#/parameters/orderByParam"
      responses:
        200:
          description: OK. Response will contain id of assigned role.
          schema:
            type: object
            properties:
              id:
                type: string
              version:
                type: string
              result:
                type: object
                properties:
                  success:
                    type: boolean
                  status:
                    type: integer
                  metadata:
                    type: object
                  content:
                    type: object
                    properties:
                      id:
                        type: string
        400:
          description: Invalid input.
          schema:
            $ref: "#/definitions/ErrorModel"
        401:
          description: If user not authenticated or does not have admin role.
          schema:
            $ref: "#/definitions/ErrorModel"
        403:
          description: If user does not have admin role.
          schema:
            $ref: "#/definitions/ErrorModel"
        500:
          description: Internal server error.
          schema:
            $ref: "#/definitions/ErrorModel"
  /roles/{roleId}/hasrole:
    get:
      description: |
        Check whether a subject has role
        It returns with a role if user has a role specified by id
      tags:
      - roles
      operationId: RoleResource.hasRole
      security:
        - bearer: []
      produces:
      - application/json
      parameters:
      - name: roleId
        in: path
        required: true
        type: string
        description: role id
      - name: filter
        in: query
        required: false
        type: string
        description: |
          fieldName=fieldValue (enum[string],optional)  - parameter for
          filtering the fields according to v3 spec

          + id
          + createdAt
          + createdBy
          + updatedAt
          + updatedBy
          + roleName
          + subjectID
      - name: fields
        in: query
        required: false
        type: string
        description: |
          fields=fieldName1,fieldName2,...,fieldN  - parameter for
          choosing which fields of role that will be included in response.

          + id
          + createdAt
          + createdBy
          + updatedAt
          + updatedBy
          + roleName
          + subjects
      - $ref: "#/parameters/offsetParam"
      - $ref: "#/parameters/limitParam"
      - $ref: "#/parameters/orderByParam"
      responses:
        200:
          description: OK
          schema:
            type: object
            properties:
              id:
                type: string
              version:
                type: string
              result:
                type: object
                properties:
                  success:
                    type: boolean
                  status:
                    type: integer
                  metadata:
                    type: object
                  content:
                    $ref: "#/definitions/RoleResponse"
        400:
          description: Invalid input.
          schema:
            $ref: "#/definitions/ErrorModel"
        401:
          description: If user not authenticated or does not have admin role.
          schema:
            $ref: "#/definitions/ErrorModel"
        404:
          description: If the role can not be found
          schema:
            $ref: "#/definitions/ErrorModel"
        500:
          description: Internal server error.
          schema:
            $ref: "#/definitions/ErrorModel"
  /roles/{roleId}/deassign:
    delete:
      description: Remove a role from user
      tags:
      - roles
      operationId: RoleResource.deassignRole
      security:
        - bearer: []
      produces:
      - application/json
      parameters:
      - name: roleId
        in: path
        required: true
        type: string
        description: role id
      - name: filter
        in: query
        required: false
        type: string
        description: |
          fieldName=fieldValue (enum[string],optional)  - parameter for
          filtering the fields according to v3 spec

          + id
          + createdAt
          + createdBy
          + updatedAt
          + updatedBy
          + roleName
          + subjectID
      - name: fields
        in: query
        required: false
        type: string
        description: |
          fields=fieldName1,fieldName2,...,fieldN  - parameter for
          choosing which fields of role that will be included in response.

          + id
          + createdAt
          + createdBy
          + updatedAt
          + updatedBy
          + roleName
          + subjects
      - $ref: "#/parameters/offsetParam"
      - $ref: "#/parameters/limitParam"
      - $ref: "#/parameters/orderByParam"
      responses:
        200:
          description: OK. Response will contain id of deassigned role.
          schema:
            type: object
            properties:
              id:
                type: string
              version:
                type: string
              result:
                type: object
                properties:
                  success:
                    type: boolean
                  status:
                    type: integer
                  metadata:
                    type: object
                  content:
                    type: object
                    properties:
                      id:
                        type: string
        400:
          description: Invalid input.
          schema:
            $ref: "#/definitions/ErrorModel"
        401:
          description: If user not authenticated or does not have admin role.
          schema:
            $ref: "#/definitions/ErrorModel"
        403:
          description: If user does not have admin role.
          schema:
            $ref: "#/definitions/ErrorModel"
        500:
          description: Internal server error.
          schema:
            $ref: "#/definitions/ErrorModel"
  /roles:
    get:
      description: Fetch all roles in the system
      tags:
      - roles
      operationId: RoleResource.getObjects
      security:
        - bearer: []
      produces:
      - application/json
      parameters:
      - name: filter
        in: query
        required: false
        type: string
        description: |
          fieldName=fieldValue (enum[string],optional)  - parameter for
          filtering the fields according to v3 spec

          + id
          + createdAt
          + createdBy
          + updatedAt
          + updatedBy
          + roleName
          + subjectID
      - name: fields
        in: query
        required: false
        type: string
        description: |
          fields=fieldName1,fieldName2,...,fieldN  - parameter for
          choosing which fields of role that will be included in response.

          + id
          + createdAt
          + createdBy
          + updatedAt
          + updatedBy
          + roleName
          + subjects
      - $ref: "#/parameters/offsetParam"
      - $ref: "#/parameters/limitParam"
      - $ref: "#/parameters/orderByParam"
      responses:
        200:
          description: OK
          schema:
            type: object
            properties:
              id:
                type: string
              version:
                type: string
              result:
                type: object
                properties:
                  success:
                    type: boolean
                  status:
                    type: integer
                  metadata:
                    type: object
                  content:
                    type: array
                    items:
                      $ref: "#/definitions/RoleResponse"
        400:
          description: Invalid input.
          schema:
            $ref: "#/definitions/ErrorModel"
        401:
          description: If user not authenticated.
          schema:
            $ref: "#/definitions/ErrorModel"
        403:
          description: If user does not have admin role.
          schema:
            $ref: "#/definitions/ErrorModel"
        404:
          description: If none is found.
          schema:
            $ref: "#/definitions/ErrorModel"
        500:
          description: Internal server error.
          schema:
            $ref: "#/definitions/ErrorModel"
    post:
      description: Creation of new role
      tags:
      - roles
      operationId: RoleResource.createObject
      produces:
      - application/json
      parameters:
      - in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/NewRoleBodyParam'
      responses:
        200:
          description: OK. Response will contain newly created role.
          schema:
            type: object
            properties:
              id:
                type: string
              version:
                type: string
              result:
                type: object
                properties:
                  success:
                    type: boolean
                  status:
                    type: integer
                  metadata:
                    type: object
                  content:
                    $ref: "#/definitions/RoleResponse"
        400:
          description: Invalid input.
          schema:
            $ref: "#/definitions/ErrorModel"
        401:
          description: If user not authenticated or does not have admin role.
          schema:
            $ref: "#/definitions/ErrorModel"
        403:
          description: If user does not have admin role.
          schema:
            $ref: "#/definitions/ErrorModel"
        500:
          description: Internal server error.
          schema:
            $ref: "#/definitions/ErrorModel"
...
definitions:
...
  Role:
    type: object
    properties:
      roleName:
        type: string
      subjects:
        type: array
        items:
          type: string
  NewRoleBodyParam:
    type: object
    properties:
      param:
        $ref: "#/definitions/Role"
...
  RoleResponse:
    type: object
    properties:
      id:
        type: integer
      createdAt:
        type: string
        format: date-time
      createdBy:
        type: string
        format: integer
      updatedAt:
        type: string
        format: date-time
      updatedBy:
        type: string
      roleName:
        type: string
      subjects:
        type: array
        items:
          type: string
...
  ResponseMetadata:
    title: Metadata object for a response
    type: object
    properties:
      totalCount:
        type: integer
        format: int32
        description: Total count of the objects
  ErrorModel:
    type: object
    properties:
      id:
        type: string
        description: unique id identifying the request
      version:
        type: string
      result:
        type: object
        properties:
          success:
            type: boolean
          status:
            description: http status code
            type: number
          metadata:
            $ref: '#/definitions/ResponseMetadata'
          content:
            type: object
...
parameters:
  offsetParam:
    name: offset
    description: "number of items to skip. Defaults to 0"
    in: query
    required: false
    type: integer
    format: int32
  limitParam:
    name: limit
    description: "max records to return. Defaults to 10"
    in: query
    required: false
    type: integer
    format: int32
  orderByParam:
    name: orderBy
    description: "sort items order by field."
    in: query
    required: false
    type: string